pyproject.toml
[project]
name = "prueba-tusdatos"
version = "0.1.0"
description = "Data Extraction Script"
requires-python = ">=3.12.0"
dependencies = [
    "aiohttp[speedups]",
    "backoff",
    "fastapi[all]",
    "orjson",
    "pydantic",
    "pydantic_settings",
    "tortoise-orm[asyncpg]",
]

[project.optional-dependencies]
dev = ["ruff", "mypy", "pip-tools", "pytest", "pytest-asyncio"]

[project.urls]
Repository = "https://github.com/lpardey/prueba-tusdatos"

[tool.ruff]
line-length = 119
target-version = "py312"
exclude = [".git", "__pycache__"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]
ignore = ["E231", "E501", "E203"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.12"
pretty = true
show_error_context = true
show_error_codes = true

follow_imports = "silent"
ignore_missing_imports = true

disallow_incomplete_defs = true
disallow_any_generics = true
disallow_subclassing_any = false
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_untyped_decorators = true

warn_unused_configs = true
warn_unreachable = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true

check_untyped_defs = true

no_implicit_optional = true
no_implicit_reexport = false

strict_equality = true

exclude = ['^build/.*', '^tests/.*']

[tool.pytest.ini_options]
python_files = ["tests.py", "test_*.py", "*_tests.py"]
asyncio_mode = "auto"
pythonpath = ["."]